move TestActionCommit to action_commit_test.go
